fix: 로또 발급 인터페이스를 만들어 역할 분리, Lotto/Person 클래스를 만들어 사용자와 로또의 객체를 구분, 실행시키는 Main클래스를 보기 쉽게 하기 위해 클래스 분리
