+++ src/main/java/kr/codesquad/service/CountNumber.java
@@ -1,25 +1,38 @@
 package kr.codesquad.service;
 
-import kr.codesquad.domain.Lotto;
+import kr.codesquad.domain.Person;
 
 import java.util.List;
 
 public class CountNumber {
 
-    MakeLotto makeLotto = new MakeLotto();
-
-    public Lotto collectNumberCount(Lotto lotto, List<Integer> collectNumber) { // 로또 일치 개수 구하기
-        for (int i = 0; i < lotto.getSize(); i++) {
-            collectNumberCountIndex(lotto,collectNumber,i);
+    public Integer checkNumber(List<Integer> inputLotto, List<Integer> compLotto) { //로또 맞는 개수 구하기
+        int cnt = 0;
+        for (int i = 0; i < compLotto.size(); i++) {
+            cnt = checkNumber(inputLotto, compLotto.get(i),cnt);
         }
-        return lotto;
+        return cnt;
     }
 
-    public void collectNumberCountIndex(Lotto lotto,List<Integer> collectNumber, int i) {
-        int cnt = makeLotto.checkNumber(lotto.getLotto(i), collectNumber);
-        if(cnt == 3) lotto.setCollectNum(0,lotto.getCollectNum(0)+1);
-        if(cnt == 4) lotto.setCollectNum(1,lotto.getCollectNum(1)+1);
-        if(cnt == 5) lotto.setCollectNum(2,lotto.getCollectNum(2)+1);
-        if(cnt == 6) lotto.setCollectNum(3,lotto.getCollectNum(3)+1);
+    public int checkNumber(List<Integer> inputLotto, int b,int cnt) {
+        if (inputLotto.contains(b)) {
+            cnt++;
+        }
+        return cnt;
+    }
+
+    public Person collectNumberCount(Person person, List<Integer> collectNumber) { // 로또 일치 개수 구하기
+        for (int i = 0; i < person.getSize(); i++) {
+            collectNumberCountIndex(person,collectNumber,i);
+        }
+        return person;
+    }
+
+    public void collectNumberCountIndex(Person person, List<Integer> collectNumber, int i) {
+        int cnt = checkNumber(person.getLotto(i), collectNumber);
+        if(cnt == 3) person.setCollectNum(0, person.getCollectNum(0)+1);
+        if(cnt == 4) person.setCollectNum(1, person.getCollectNum(1)+1);
+        if(cnt == 5) person.setCollectNum(2, person.getCollectNum(2)+1);
+        if(cnt == 6) person.setCollectNum(3, person.getCollectNum(3)+1);
     }
 }
