+++ src/main/java/kr/codesquad/service/MakeLottoCustom.java
@@ -1,0 +1,10 @@
+package kr.codesquad.service;
+
+import java.util.List;
+
+public class MakeLottoCustom implements MakeLottoInterface{
+    @Override
+    public List<Integer> make() {
+        return null;
+    }
+}
