+++ src/main/java/kr/codesquad/service/MakeLottoInterface.java
@@ -1,0 +1,8 @@
+package kr.codesquad.service;
+
+import java.util.List;
+
+public interface MakeLottoInterface {
+
+    List<Integer> make();
+}
