+++ src/main/java/kr/codesquad/service/ScannerInput.java
@@ -5,13 +5,7 @@ import java.util.Scanner;
 public class ScannerInput {
     Scanner sc = new Scanner(System.in);
 
-    public int moneyInput() {
-        int haveMoney = sc.nextInt();
-        return haveMoney;
-    }
-
-    public int winNumberInput() {
-        int winNumber = sc.nextInt();
-        return winNumber;
+    public int inputNumber() {
+        return sc.nextInt();
     }
 }
